About section: splash background
