+++ lib/core/core.dart
@@ -2,3 +2,4 @@ export 'extensions/extensions.dart';
 export 'helpers/helpers.dart';
 export 'theme/theme.dart';
 export 'utils/utils.dart';
+export 'widgets/widgets.dart';
