+++ lib/core/widgets/animated_container.dart
