+++ lib/core/widgets/widgets.dart
@@ -1,0 +1,4 @@
+export 'animated_button.dart';
+export 'animated_container.dart';
+export 'gradient_text.dart';
+export 'splashes.dart';
